+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project should be documented in this file.
 
 ### v0.0.1-rc1.4
-- Add Exporters to Prometheus (#41)
+- Add Exporters to Prometheus (#42)
 
 
 ### v0.0.1-beta4